Make the admin menu use the inline SVG that's then colorized instead of the font (which had rendering shortcomings)
Using a basic font (which this was set to use) lacks the opacity differentiation of elements in the icon we want, and the admin menu doesn't have enough selectors available for the icon to implement the multi-glyph icon like the admin bar does. The next best thing (in terms of rendering quality and feasibility) is to use the SVG inline.