minify debug mode - comments containing URLs were handled incorrectly
When minify debug mode enabled: single-line comments with an url were
processed incorrectly, causing removal of end-of-comment sequence

For example, this line:

/* Generated by Font Squirrel (http://www.fontsquirrel.com) on May 30,
2013 */

Gets converted to:

/* Generated by Font Squirrel (http:

So it thinks the comment is not finished.

That behaviour was breaking web rendering because large portions of
valid CSS get deactivated inside a comment when combining them inside
the big css with line comments.

Taken updated content from original minify library as a fix