dont Vary by User-Agent anymore for HTML BC caching
It was common earlier to provide user-agent specific content, so BC sets
Vary: User-Agent header by default for HTML BC caching.
And only was set for Apache, which is inconsistent.

It's not the case anymore, and it blocks FSD CDN from caching which
became popular.
So we remove that rule.