Redis: support connecting to servers using TLS with self-signed certificates (e.g., Heroku Redis) (#494)
* redis options: include mention of using 'tls://' to connect over TLS

adds on to https://github.com/W3EDGE/w3-total-cache/issues/67 to include (extremely minimal) in-plugin documentation on how to connect to redis using TLS

* redis options: include option to disable TLS certificate verification

Adds on to https://github.com/W3EDGE/w3-total-cache/issues/67 to include the option of _not_ verifying the redis server's TLS certificate when connecting over TLS.

This option is needed when connecting to a redis server configured to use TLS with self-signed certificates. In particular, Heroku Redis falls under this category.

With this new option TLS certificates continue to be verified by default, but users who need to now have the option of disabling certificate verification.