Cache_Redis: restore support for disabling TLS certificate verification
This was initially added as part of 2.2.2, but was later removed in 2.2.3 while fixing PHP warnings related to the number of args being passed to pconnect (#533).

After refactoring how Redis connection arguments are being calculated/passed, we can now support disabling TLS certificate verification when the installed version of phpredis supports it.

phpredis added initial support for the stream context in version 5.3.0 (https://github.com/phpredis/phpredis/issues/1600) and support for RedisCluster in 5.3.2.