Fix Memcached test using SASL auth
Fixes #448. When testing a Memcached cache, in addition to the configured servers, also pass binary protocol, username and password. This allows the test to be run correctly when SASL authentication is used (when a username and password are supplied). The binary protocol option is also passed as SASL auth with Memcached requires the binary protocol to be used.