Use SimpleXMLElement to parse sitemaps and RSS feeds.
The existing code relies on regex, which can be brittle. Users are reporting that the use of valid XML CDATA tags in sitemaps are causing W3 Total Cache to load the wrong URLs.

Examples of problem reports:
* https://wordpress.org/support/topic/sitemap-precachcing-is-pulling-wrong-urls/
* https://wordpress.org/support/topic/page-cache-pre-load-wrong-urls-fetched/

This code changes the plugin to use SimpleXMLElement to handle the XML parsing, instead of regex.